Allocators fix, StringArena fix + range-for, ls(), load_assets_dir()
LinkedListAllocator::try_split() was not updating _used variable
correctly when split node was not free, now its ok.

StringArena uses offsets instead of StringViews since it would be
rather problematic as it can grow and reallocate its memory.
Also Iterator and begin(), end() were added to allow for range-for.

platform::ls() takes dirname and returns names of files in the given
directory as StringArena.

serialization::load_assets_dir() is one of the first to use ls(), it
takes dirname and attempts to load all .past asset files.